requirements.txt
alembic==1.8.1
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Flask==2.1.3
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.1
